Adding comments on ways to change this to accomodate much larger numbers.

+++ math/totient/phi.h
@@ -4,14 +4,15 @@
 #include <stdint.h>
 #include <stdio.h>
 
-#define SQUARE_SZ 1025
+#define SQUARE_SZ 1025 /* This can be increased to 65000 */
 #define PRIME_SIZE 1048576
-#define SMALL_PRIME_SZ 82025
+#define SMALL_PRIME_SZ 82025 
 
 /*
  * Boolean array declaring a number prime or not
  */
-int32_t prime_array[PRIME_SIZE];
+//int32_t prime_array[PRIME_SIZE];
+uint8_t prime_array[PRIME_SIZE];
 
 /*
  * An array of the squares of integers
@@ -19,13 +20,16 @@ int32_t prime_array[PRIME_SIZE];
 uint32_t square_int[SQUARE_SZ];
 
 /*
- * A list of small primes
+ * A list of small primes.  This can be used 
+ * quickly determine if a number greater than 
+ * 1048576, but less than 1048576**2.
  */
 uint32_t small_primes[SMALL_PRIME_SZ];
 
 /*
  * Computes the integer part of the 
- * square root of n.
+ * square root of n.  This can easily
+ * be extended to much larger numbers.
  */
 uint32_t int_sqrt ( uint32_t n );
